requirements.txt
azure-cognitiveservices-speech==1.37.0
blinker==1.7.0
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
pydub==0.25.1
python-dotenv==1.0.1
Werkzeug==3.0.2
